feat: 좋아요 버퍼 동기화를 위한 Graceful Shutdown 구현 (#26)
- 서버 종료(SIGTERM) 시 인메모리 버퍼 내 잔량 데이터 유실 방지
- LikeSyncScheduler 로직을 LikeSyncService로 분리하여 SRP 준수
- @PreDestroy 훅을 통한 종료 직전 강제 Flush 메커니즘 구축
- application.yml 내 graceful shutdown 설정 적용 (30초 대기)
